requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests>=2.28.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
Pillow==10.0.0
python-dotenv==1.1.1
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.0
tensorflow==2.10.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0